+++ kroxylicious-operator/src/main/java/io/kroxylicious/kubernetes/operator/reconciler/kafkaproxy/ClusterRouteDependentResource.java
@@ -1,0 +1,74 @@
+/*
+ * Copyright Kroxylicious Authors.
+ *
+ * Licensed under the Apache Software License version 2.0, available at http://www.apache.org/licenses/LICENSE-2.0
+ */
+package io.kroxylicious.kubernetes.operator.reconciler.kafkaproxy;
+
+import java.util.Map;
+import java.util.Set;
+
+import io.fabric8.openshift.api.model.Route;
+import io.javaoperatorsdk.operator.api.reconciler.Context;
+import io.javaoperatorsdk.operator.processing.dependent.BulkDependentResource;
+import io.javaoperatorsdk.operator.processing.dependent.kubernetes.BooleanWithUndefined;
+import io.javaoperatorsdk.operator.processing.dependent.kubernetes.CRUDKubernetesDependentResource;
+import io.javaoperatorsdk.operator.processing.dependent.kubernetes.KubernetesDependent;
+import io.javaoperatorsdk.operator.processing.event.NoEventSourceForClassException;
+
+import io.kroxylicious.kubernetes.api.v1alpha1.KafkaProxy;
+import io.kroxylicious.kubernetes.operator.model.networking.ProxyNetworkingModel;
+import io.kroxylicious.kubernetes.operator.resolver.ClusterResolutionResult;
+
+import static io.kroxylicious.kubernetes.operator.ResourcesUtil.toByNameMap;
+
+/**
+ * Generates the OpenShift {@code Route} for a single virtual cluster.
+ */
+@KubernetesDependent(useSSA = BooleanWithUndefined.TRUE)
+public class ClusterRouteDependentResource
+        extends CRUDKubernetesDependentResource<Route, KafkaProxy>
+        implements BulkDependentResource<Route, KafkaProxy, String> {
+
+    public ClusterRouteDependentResource() {
+        super(Route.class);
+    }
+
+    @Override
+    public Map<String, Route> desiredResources(
+                                               KafkaProxy primary,
+                                               Context<KafkaProxy> context) {
+        KafkaProxyContext kafkaProxyContext = KafkaProxyContext.proxyContext(context);
+        var model = kafkaProxyContext.model();
+        var clusterNetworkingModels = model.clustersWithValidNetworking().stream()
+                .map(ClusterResolutionResult::cluster)
+                .filter(cluster -> !kafkaProxyContext.isBroken(cluster))
+                .flatMap(cluster -> model.networkingModel().clusterIngressModel(cluster).stream())
+                .toList();
+
+        var routeStream = clusterNetworkingModels.stream()
+                .flatMap(ProxyNetworkingModel.ClusterNetworkingModel::routes);
+
+        return routeStream.collect(toByNameMap());
+    }
+
+    @Override
+    public Map<String, Route> getSecondaryResources(
+                                                    KafkaProxy primary,
+                                                    Context<KafkaProxy> context) {
+        // This try/catch is a temporary workaround until the following bug is fixed: https://github.com/operator-framework/java-operator-sdk/issues/3249
+        try {
+            Set<Route> secondaryResources = context.eventSourceRetriever().getEventSourceFor(Route.class)
+                    .getSecondaryResources(primary);
+            return secondaryResources.stream().collect(toByNameMap());
+        }
+        catch (NoEventSourceForClassException e) {
+            return Map.of();
+        }
+    }
+
+    @Override
+    public void deleteTargetResource(KafkaProxy primary, Route resource, String key, Context<KafkaProxy> context) {
+        context.getClient().resource(resource).delete();
+    }
+}